pyproject.toml
[project]
name = "media-kb-mcp"
version = "0.1.0"
description = "Media Knowledge Base MCP - Semantic search over video transcripts and other media content"
requires-python = ">=3.10"

dependencies = [
    "numpy",
    "pandas",
    "openai",
    "langchain",
    "transformers",
    "datasets",
    "huggingface_hub",
    "python-dotenv",
    "wikipedia",
    "langchain-openai",
    "yt-dlp>=2025.5",
    "mcp[cli]",
    "langchain-community",
    "faiss-cpu",
    "scikit-learn",
    "pyarrow",
    "httpx",
    "psycopg2-binary",
    "ipykernel",
    "jupyterlab",
    "langchain-ollama",
    "langchain-huggingface",
    "sentence-transformers",
    "lancedb>=0.17.0",
    "voyageai>=0.3.0",
    "pydantic>=2.0.0",
]

[dependency-groups]
dev = [
    "ipykernel>=6.29.5",
    "jupyter>=1.1.1",
]
